navbar.tsx moved out of layouts folder and into components
layouts folder deleted as there is no use for it
